requirements.txt
httpx>=0.27,<1.0
pyyaml>=6.0
jinja2>=3.1
click>=8.1
rich>=13.0
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
itsdangerous>=2.1
apscheduler>=3.10
bcrypt>=4.0
python-dotenv>=1.0
